+++ packages/helix-shared-config/src/MountConfig.js
@@ -87,6 +87,7 @@ const githubDecorator = {
       ref,
       path: m.path,
       basePath: path,
+      url: m.url,
     };
     return ret;
   },