+++ packages/helix-shared-config/src/index.js
@@ -17,6 +17,9 @@ const Strain = require('./Strain.js');
 const MarkupConfig = require('./MarkupConfig');
 const Condition = require('./Condition.js');
 const { optionalConfig, requiredConfig } = require('./config-wrapper');
+const DataEmbedValidator = require('./DataEmbedValidator.js');
+const ValidationError = require('./ValidationError.js');
+const IgnoreConfig = require('./IgnoreConfig.js');
 
 module.exports = {
   HelixConfig,
@@ -25,7 +28,10 @@ module.exports = {
   RedirectConfig,
   Strain,
   MarkupConfig,
+  IgnoreConfig,
   Condition,
   optionalConfig,
   requiredConfig,
+  DataEmbedValidator,
+  ValidationError,
 };